requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community


# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

